requirements_network.txt
# Network-based remote inference requirements
# For both Raspberry Pi (client) and Laptop (server)

# Core networking and data processing
numpy>=1.21.0
opencv-python>=4.5.0
pyserial>=3.5

# Server-side ML requirements (laptop)
torch>=1.9.0
torchvision>=0.10.0

# Data handling
pandas>=1.3.0
matplotlib>=3.5.0  # For visualization/debugging

# Optional: For advanced networking features
# websockets>=10.0  # Alternative to raw sockets
# msgpack>=1.0.0    # More efficient serialization than JSON